# Conflicts:
#	src/Nethermind/Nethermind.Consensus.AuRa/Transactions/TxCertifierFilter.cs
#	src/Nethermind/Nethermind.Consensus.AuRa/Transactions/TxPermissionFilter.cs

+++ src/Nethermind/Nethermind.Baseline.Plugin/VaultPlugin.cs
@@ -57,7 +57,7 @@ namespace Nethermind.Plugin.Baseline
 
                 // TODO: change vault to provide, use sealer to set the gas price as well
                 // TODO: need to verify the timing of initializations so the TxSender replacement works fine
-                _api.TxSender = new VaultTxSender(vaultSigner, _vaultConfig, _api.ChainSpec.ChainId);
+                _api.TxSender = new VaultTxSender(vaultSigner, _vaultConfig, (int)_api.ChainSpec.ChainId);
             }
         }
